requirements.txt
GitPython 
networkx 
graphviz
equests 
  streamlit pygments
  r